Move code into separate files in the library
group code for artists and albums separately

this way multiple tools may access this code.
