+++ Sources/iTunes/SortableName+Similar.swift
@@ -1,0 +1,16 @@
+//
+//  SortableName+Similar.swift
+//  itunes_json
+//
+//  Created by Greg Bolsinga on 12/1/24.
+//
+
+extension SortableName: Similar {
+  var cullable: Bool {
+    sorted.isEmpty
+  }
+
+  func isSimilar(to other: SortableName) -> Bool {
+    self.name.isSimilar(to: other.name)
+  }
+}
